requirements.txt
groq
aiogram
python-dotenv
redis
httpx
openai
google-generativeai
pylovepdf